# docs/swagger.yaml
basePath: /
definitions:
...
  api.AuthorityIdInfo:
    properties:
      authorityId:
        type: string
    type: object
...
  api.GetById:
    properties:
      id:
        type: number
    type: object
  api.IdInfo:
    properties:
      id:
        type: number
    type: object
  api.RegistAndLoginStuct:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  editorapi.examinParam:
    properties:
      catalog_uuid:
        description: 目录UUID
        type: string
      comment:
        description: 审核评语
        type: string
      examin_state:
        description: 审核状态，0正在编辑，1提交审核，2审核通过，3审核没通过
        type: integer
      user_uuid:
        description: 分配的权限UUID
        type: string
    type: object
  editorapi.examinSubmitParam:
    properties:
      catalog_uuid:
        description: 目录UUID
        type: string
    type: object
...
  modelInterface.PageInfo:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
    type: object
...
  sysModel.SysBaseMenu:
    properties:
      children:
        items:
          $ref: '#/definitions/sysModel.SysBaseMenu'
        type: array
      component:
        type: string
      hidden:
        type: boolean
      icon:
        type: string
      name:
        type: string
      nickName:
        type: string
      parentId:
        type: string
      path:
        type: string
      sort:
        type: string
      title:
        type: string
    type: object
...
  sysModel.SysWorkflow:
    properties:
      workflowDescription:
        description: 工作流描述
        type: string
      workflowName:
        description: 工作流英文id
        type: string
      workflowNickName:
        description: 工作流名称
        type: string
      workflowStep:
        description: 工作流步骤
        items:
          $ref: '#/definitions/sysModel.SysWorkflowStepInfo'
        type: array
    type: object
  sysModel.SysWorkflowStepInfo:
    properties:
      isEnd:
        description: 是否是完结流节点
        type: boolean
      isStrat:
        description: 是否是开始流节点
        type: boolean
      stepAuthorityID:
        description: 操作者级别id
        type: string
      stepName:
        description: 工作流名称
        type: string
      stepNo:
        description: 步骤id （第几步）
        type: number
      workflowID:
        description: 所属工作流ID
        type: integer
    type: object
info:
  contact: {}
  description: 全球说编辑器后台API服务文档
  license: {}
  title: 全球说编辑器后台API服务文档
  version: 0.0.1
paths:
...
  /editor/examin:
    post:
      consumes:
      - application/json
      parameters:
      - description: '内容审核: 审核状态，0正在编辑，1提交审核，2审核通过，3审核没通过'
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.examinParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 审核接口
      tags:
      - EditorCatalogAPI(目录接口)
  /editor/examin/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审核状态重置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.examinSubmitParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 审核状态重置
      tags:
      - EditorCatalogAPI(目录接口)
  /editor/examin/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 提交审核
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.examinSubmitParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 提交审核
      tags:
      - EditorCatalogAPI(目录接口)
...
  /menu/addBaseMenu:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新增菜单
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sysModel.SysBaseMenu'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增菜单
      tags:
      - menu
  /menu/addMenuAuthority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 增加menu和角色关联关系
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.AuthorityIdInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取指定角色menu
      tags:
      - authorityAndMenu
  /menu/deleteBaseMenu:
    post:
      consumes:
      - application/json
      parameters:
      - description: 删除菜单
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.IdInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除菜单
      tags:
      - menu
  /menu/getBaseMenuById:
    post:
      consumes:
      - application/json
      parameters:
      - description: 根据id获取菜单
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取菜单
      tags:
      - menu
  /menu/getBaseMenuTree:
    post:
      parameters:
      - description: 可以什么都不填
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.RegistAndLoginStuct'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"返回成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取用户动态路由
      tags:
      - authorityAndMenu
  /menu/getMenu:
    post:
      parameters:
      - description: 可以什么都不填
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.RegistAndLoginStuct'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"返回成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取用户动态路由
      tags:
      - authorityAndMenu
  /menu/getMenuList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页获取基础menu列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelInterface.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取基础menu列表
      tags:
      - menu
  /menu/updataBaseMen:
    post:
      consumes:
      - application/json
      parameters:
      - description: 更新菜单
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sysModel.SysBaseMenu'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新菜单
      tags:
      - menu
...
  /workflow/createWorkFlow:
    post:
      parameters:
      - description: 注册工作流接口
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sysModel.SysWorkflow'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"注册成功"}'
          schema:
            type: string
      summary: 注册工作流
      tags:
      - workflow
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"